Implementing preset stuff
Added preset functinalities:
- completed: load random preset
- added: save preset (user)
- added: load preset (user)
- added: add preset(user)
- added: load preset folder (user)

Implemented folder structure for presets:
Presets are stored in a folder with the same name as the sonorium
preset file.
The sonorium preset file is contained in this folder.
As well as all output presets.
If the sonorium preset name is "sonorium_presets.json", output presets
names will be:
01_sonorium_presets.json
02_sonorium_presets.json
03_sonorium_presets.json
etc……

Note: There are several bugs coming across these implementations. They
should all be verified when all the preset stuff is implemented. Plus,
the 2D matrix is not loading correctly when changing presets